python/examples/wasd_server/requirements.txt
-f ../../../prebuilt
bosdyn-client >= 3.1

Flask
Pillow
